hlas/requirements.txt
fastapi
uvicorn
python-dotenv
langchain_openai
langchain
crewai
weaviate-client
pyyaml